Web API/requirements.txt

Flask==1.1.2
Flask-Elasticsearch==0.2.5
Flask-Ext==0.1

gunicorn==20.0.4

Werkzeug==1.0.1

